[FIX] 필요없는 필드 삭제 및 삭제 메서드 추가
