+++ src/main/java/com/wearei/finalsamplecode/domain/board/dto/request/BoardUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.wearei.finalsamplecode.domain.board.dto;
+package com.wearei.finalsamplecode.domain.board.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,12 +8,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BoardUpdateRequestDto {
-
     private Long teamId;
     private Long id;
     private String title;
     private String contents;
-    private String backgroundImage;
-    private int likes;
-    private boolean isDeleted;
 }
