+++ src/main/java/com/wearei/finalsamplecode/domain/comment/dto/request/CommentCreateRequestDto.java
@@ -7,5 +7,4 @@ public class CommentCreateRequestDto {
     private Long teamId;
     private Long boardId;
     private String contents;
-    private boolean isDeleted;
 }
